requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.1.1
supabase==2.18.1
pydantic>=2.11.7,<3.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
